Continue to extract resolveFns() et al from functionResolution.cpp
